PagBackend v1.0.0 PagBackend.Api

Helper com funcoes gerais para auxiliar na construção dos modulos.

Functions

check_halted_conn(conn, fun)

Checa por um halt em uma requisicao que foi invalidada.

get_authorization_header(conn)

Verifica para o header de Authorization e Splita o Bearer retornando a token.

handle_error(op, fun)

Checa para erros na requisição, agrupa os erros e invoca o callback passado.

handle_success(op, fun)

Checa para resposta a requisições bem sucedidas

traverse_errors_(errors)

Atravessa os erros gerados pelo changeset substituindo as variaves.