PagBackend v1.0.0 PagBackend.Api.AWS

Esse modulo contem o dominio/interface das funcionalidades da AWS dentro da API

Note: Para as funcoes que interagem com o API cognito, cheque as regras de parametrizacao declaradas na criacao do pool/cliente.

Functions

aws_delete_user(params)

Delete um usuario do Cognito

aws_init_auth(params)

Inicia o processo de autenticacao do usuario. O parametro de autenticacao é USER_PASSWORD_AUTH. Por default para primeiro login com o password temporario, se o usuario estiver com o status de FORCE_CHANGE_PASSWORD essa funcao retorna um Challenge para ser requisitado a mudanca de senha para o cliente com a funcao aws_respond_challenge/1. Caso contrario retorna as devidas tokens.

aws_register_user(params)

Funcao para registrar um usuario dentro do AWS Cognito

aws_register_user_with_pass(params, pass)

Registra um usuario com o password temporario especificado

aws_respond_challenge(params)

Responde o challenge do processo de autenticacao. Para a v1 essa funcao atende apenas ao parametro NEW_PASSWORD_REQUIRED