PagBackend v1.0.0 PagBackend.Api.DAO

Database wrapper para funcoes basicas de banco de dados

Functions

all(schema, params)

Retorna todos as rows do schema passado

create(schema, mod, params \\ %{})

Cria um registro no banco de dados.

delete(schema, id)

Deleta um registro da tabela do schema.

preload(op, field)

Carrega relationships de chaves estrangeiras do schema

show(schema, id)

Atraves do schema e de um id, retorna a row no formato {:ok, row}